Report unique test ID to ABQ as test location and index in describe block
ABQ expects these test IDs to be unique - prior to this patch, they were
not necessarily, as they only contained the display name of the test. In
this patch we make the test ID unique by listing its file path, location
in the file, and its index in the parent describe block. The latter is
needed to uniquely identify tests that occur at the same location.

To deal with different paths between filesystems, make the path relative
to the root directory.