Have ABQ test ID generation be based on unique test path
To support focused test retries, we'll need circus to execute tests
based on a passed unique ID. The current strategy of using location for
unique identification is not enough because we don't necessarily know a
test's location until after it executes. Instead, use the path of the
test through describe blocks, which must be unique.